s/model/completion
`@text.model` does not convey what the function is doing,
`text.completion` does a much better job at this.
