+++ outlines/__init__.py
@@ -31,7 +31,7 @@ References
 from outlines.function import fn
 from outlines.image import as_image
 from outlines.program import chain, program
-from outlines.text import as_string, render
+from outlines.text import as_string, completion, render
 
 __all__ = [
     "chain",
